web-animations/animation-player/animation-player-start-time-test.html

<!DOCTYPE html>
<title>Animation player startTime attribute test</title>
<meta name="assert" content="Returns the start time of this player. Setting this attribute updates the player start time using the procedure defined in section 3.5.2.6 Updating the player start time.">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationPlayer-startTime">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
function newAnimationPlayer(doc) {
    var animationTarget = doc.createElement('div');
    doc.body.appendChild(animationTarget);
    var animation = new Animation(animationTarget, [], 5);
    return doc.timeline.play(animation);
}


test(function() {
    var doc = newHTMLDocument();
    var player = newAnimationPlayer(doc);

    assert_not_equals(player.startTime, null,
        'The value of AnimationPlayer.startTime attribute should not be of type null');
}, 'The value of AnimationPlayer.startTime attribute should not be null');


test(function() {
    var doc = newHTMLDocument();
    var player = newAnimationPlayer(doc);

    assert_throws('TypeError',function () {
        player.startTime = [];
    }, 'TypeError should be thrown if AnimationPlayer.startTime attribute ' +
        'is set to not double value');
}, 'Test that TypeError is thrown if AnimationPlayer.startTime attribute ' +
    'is set to not double value');


test(function() {
    var doc = newHTMLDocument();
    var player = newAnimationPlayer(doc);

    var newStartTime = player.startTime + 100;
    player.startTime = newStartTime;

    assert_equals(player.startTime, newStartTime,
        'The value of AnimationPlayer.startTime attribute should be set to the new value');
	// TODO Is there a way to test the following "2. Set the hold time to null."?
}, 'Test setting value of AnimationPlayer.startTime attribute. Player is not paused');


test(function() {
    var doc = newHTMLDocument();
    var player = newAnimationPlayer(doc);

    var newStartTime = player.startTime + 100;
    player.pause();
    player.startTime = newStartTime;

    var expectedTime = player.timeline.currentTime - newStartTime * player.playbackRate;
    assert_equals(player.startTime, expectedTime,
        'The value of AnimationPlayer.startTime attribute should be set to ' +
        'hod time value');
}, 'Test setting value of AnimationPlayer.startTime attribute. Player is paused');
</script>
</body>
